+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Changelog
 
-No releases yet.
+## v0.2.0 (2022-09-24) _unreleased_
+
+- Added `bufferedRandom`, which calls `crypto.getRandomBytes` in blocks instead of on every UUID generation call, to improve performance significantly.
+- Changed the signature of `EntropySource` to accept the number of desired random bytes.
+
+## v0.1.0 (2022-09-23)
+
+- Initial release
+- Wrote the [readme](./README.md)